+++ app/controllers/passwordless/sessions_controller.rb
@@ -105,14 +105,12 @@ module Passwordless
     end
 
     def find_authenticatable
-      email = params[:passwordless][email_field].downcase
+      email = params[:passwordless][email_field].downcase.strip
 
       if authenticatable_class.respond_to?(:fetch_resource_for_passwordless)
         authenticatable_class.fetch_resource_for_passwordless(email)
       else
-        authenticatable_class.where(
-          "lower(#{email_field}) = ?", params[:passwordless][email_field].downcase
-        ).first
+        authenticatable_class.where("lower(#{email_field}) = ?", email).first
       end
     end